// src/types/meal.ts
export interface Meal {
  id: string;
  date: Date;
  photoDataUri: string;
  description: string;
  estimatedCalories: number | null;
  protein: number | null;
  carbs: number | null;
  fat: number | null;
  notes?: string;
  mealType: 'Breakfast' | 'Lunch' | 'Dinner' | 'Snack' | 'Pre-workout' | 'Post-workout';
  recognizedItems?: string[] | null; // Added field for recognized items
  created: Date; // Timestamp when the meal was created
  updated: Date; // Timestamp when the meal was last updated
}

/**
 * Type for the meal data needed to create a new meal entry or update an existing one.
 * Omits the id which will be generated by Firestore
 */
export type AddMealDTO = Omit<Meal, 'id' | 'created' | 'updated'>;

export type UpdateMealDTO = Omit<Meal, 'created' | 'updated'>;

export const mealTypes: Meal['mealType'][] = [
  'Breakfast',
  'Lunch',
  'Dinner',
  'Snack',
  'Pre-workout',
  'Post-workout',
];

export const estimationTypes = [
  { value: 'calories_macros', label: 'Calories & Macros' },
  { value: 'calories_only', label: 'Calories Only' },
  { value: 'macros_only', label: 'Macros Only' },
] as const;

export type EstimationType = (typeof estimationTypes)[number]['value'];